requirements.txt
Flask==2.2.2
pyodbc==4.0.35
python-dotenv==1.0.1
flake8==7.0.0
nose